+++ .idea/modules/ThingiMarket.main.iml
@@ -1,8 +1,8 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <module version="4">
   <component name="AdditionalModuleElements">
-    <content url="file://$MODULE_DIR$/../../ThingiMarket/build/generated/sources/annotationProcessor/java/main">
-      <sourceFolder url="file://$MODULE_DIR$/../../ThingiMarket/build/generated/sources/annotationProcessor/java/main" isTestSource="false" generated="true" />
+    <content url="file://$MODULE_DIR$/../../MyeongjiMarket/build/generated/sources/annotationProcessor/java/main">
+      <sourceFolder url="file://$MODULE_DIR$/../../MyeongjiMarket/build/generated/sources/annotationProcessor/java/main" isTestSource="false" generated="true" />
     </content>
   </component>
 </module>
