+++ MyeongjiMarket/src/main/java/com/example/myeongjimarket/MyeongjiMarketApplication.java
@@ -1,0 +1,13 @@
+package com.example.myeongjimarket;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class MyeongjiMarketApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(MyeongjiMarketApplication.class, args);
+	}
+
+}
