+++ MyeongjiMarket/src/main/java/com/example/myeongjimarket/domain/Post/controller/PostController.java
@@ -1,0 +1,49 @@
+package com.example.myeongjimarket.domain.Post.controller;
+
+import com.example.myeongjimarket.domain.Post.Post;
+import com.example.myeongjimarket.domain.Post.service.PostService;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Controller;
+import org.springframework.ui.Model;
+import org.springframework.web.bind.annotation.*;
+
+@Controller
+@RequestMapping("/posts")
+public class PostController {
+    private final PostService postService;
+    @Autowired
+    public PostController(PostService postService) {
+        this.postService = postService;
+    }
+    @GetMapping
+    public String listPosts(Model model) {
+        model.addAttribute("posts", postService.getAllPosts());
+        return "posts/index";
+    }
+    @GetMapping("/new")
+    public String newPostForm(Model model) {
+        model.addAttribute("post", new Post());
+        return "posts/new";
+    }
+    @PostMapping
+    public String savePost(@ModelAttribute Post post) {
+        postService.savePost(post);
+        return "redirect:/posts";
+    }
+    @GetMapping("/{id}")
+    public String editPostForm(@PathVariable Long id, Model model) {
+        model.addAttribute("post", postService.getPostById(id));
+        return "posts/edit";
+    }
+    @PostMapping("/{id}")
+    public String updatePost(@PathVariable Long id, @ModelAttribute Post post) {
+        post.setId(id);
+        postService.savePost(post);
+        return "redirect:/posts";
+    }
+    @PostMapping("/delete/{id}")
+    public String deletePost(@PathVariable Long id) {
+        postService.deletePost(id);
+        return "redirect:/posts";
+    }
+}
