+++ MyeongjiMarket/src/main/java/com/example/myeongjimarket/domain/Post/dto/PostDto.java
@@ -1,0 +1,17 @@
+package com.example.myeongjimarket.domain.Post.dto;
+
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+@Getter
+@Setter
+@NoArgsConstructor
+@AllArgsConstructor
+public class PostDto {
+        private Long id;
+        private String title;
+        private String content;
+}
