+++ MyeongjiMarket/src/main/java/com/example/myeongjimarket/domain/Post/service/PostService.java
@@ -1,0 +1,33 @@
+package com.example.myeongjimarket.domain.Post.service;
+
+import com.example.myeongjimarket.domain.Post.Post;
+import com.example.myeongjimarket.domain.Post.repository.PostRepository;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Service;
+import java.util.List;
+
+@Service
+public class PostService {
+    private final PostRepository postRepository;
+
+    @Autowired
+    public PostService(PostRepository postRepository) {
+        this.postRepository = postRepository;
+    }
+
+    public List<Post> getAllPosts() {
+        return postRepository.findAll();
+    }
+
+    public Post getPostById(Long id) {
+        return postRepository.findById(id).orElse(null);
+    }
+
+    public Post savePost(Post post) {
+        return postRepository.save(post);
+    }
+
+    public void deletePost(Long id) {
+        postRepository.deleteById(id);
+    }
+}
