+++ MyeongjiMarket/src/test/java/com/example/myeongjimarket/MyeongjiMarketApplicationTests.java
@@ -1,0 +1,13 @@
+package com.example.myeongjimarket;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class MyeongjiMarketApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
