--- ThingiMarket/src/main/java/com/example/thingimarket/ThingiMarketApplication.java
@@ -1,13 +1,0 @@
-package com.example.thingimarket;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class ThingiMarketApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(ThingiMarketApplication.class, args);
-    }
-
-}
