--- ThingiMarket/src/main/java/com/example/thingimarket/domain/transaction/TransactionPost.java
@@ -1,33 +1,0 @@
-package com.example.thingimarket.domain.transaction;
-
-import jakarta.persistence.*;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-import java.math.BigDecimal;
-import java.util.Date;
-
-@Getter
-@Setter
-@AllArgsConstructor
-@NoArgsConstructor
-@Entity
-public class TransactionPost {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-    private String title;
-    private String content;
-    private String topic;
-    @Lob
-    private byte[] picture;
-    private BigDecimal price;
-    private String place;
-    private Boolean deadline;
-    @Temporal(TemporalType.TIMESTAMP)
-    private Date createdAt;
-    @Temporal(TemporalType.TIMESTAMP)
-    private Date updatedAt;
-}
