--- ThingiMarket/src/main/java/com/example/thingimarket/dto/post/PostDto.java
@@ -1,17 +1,0 @@
-package com.example.thingimarket.dto.post;
-
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-@Getter
-@Setter
-@NoArgsConstructor
-@AllArgsConstructor
-public class PostDto {
-        private Long id;
-        private String title;
-        private String content;
-}
