--- ThingiMarket/src/main/resources/application-seohyun.yml
@@ -1,18 +1,0 @@
-
-spring:
-  config:
-    activate:
-      on-profile: seohyun
-  datasource:
-    url: "?"
-    username: "?"
-    password: "?"
-    driver-class-name: com.mysql.cj.jdbc.Driver
-  jpa:
-    hibernate:
-      ddl-auto: none
-    properties:
-      hibernate:
-        show_sql: true
-        format_sql: true
-        dialect: org.hibernate.dialect.MySQL8Dialect
