--- ThingiMarket/src/test/java/com/example/thingimarket/ThingiMarketApplicationTests.java
@@ -1,13 +1,0 @@
-package com.example.thingimarket;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class ThingiMarketApplicationTests {
-
-    @Test
-    void contextLoads() {
-    }
-
-}
